refactor: expose event-bus via core/event-bus re-export barrel
Adds src/core/event-bus.ts as a stable import path for the event-bus
so cross-cutting subsystems (adapters, consensus, pipeline) can wire
into it without crossing the agents-layer boundary flagged by the
fitness-audit layerSeparation check.

- New: src/core/event-bus.ts re-exports public surface from existing
  implementation in src/agents/collaboration/event-bus*
- Migrated: resilient-adapter and weighted-voting now import from
  the new core path
- No breaking change: implementation stays put; 35+ existing importers
  at old path continue to work
- Physical move is v3.0-gated (see #2066) because internal
  event-bus-events.ts has coupling to collaboration-types.ts that
  must be decoupled first

Fitness: 99 -> 100 (layer-separation violations 2 -> 0).
